Ajout token delete category; menu backoffice; remplacement image url absole par image uploadée; ajout pages erreur

+++ app/src/Controller/Admin/CategoryController.php
@@ -9,6 +9,8 @@ use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Csrf\CsrfToken;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 #[Route('/admin/category', name: 'admin_category_')]
@@ -65,12 +67,18 @@ class CategoryController extends AbstractController
     }
 
     #[Route('/delete/{id}', name: 'delete')]
-    public function delete(Category $category, ManagerRegistry $doctrine): Response
+    public function delete(Category $category, ManagerRegistry $doctrine, Request $request, CsrfTokenManagerInterface $csrfTokenManager): Response
     {
-        $em = $doctrine->getManager();
-        $em->remove($category);
-        $em->flush();
-        //$this->addFlash('success', 'Article supprimé !');
+        $token = new CsrfToken('delete', $request->query->get('_csrf_token'));
+        if ($csrfTokenManager->isTokenValid($token)) 
+        {
+            $em = $doctrine->getManager();
+            $em->remove($category);
+            $em->flush();
+            $this->addFlash('success', 'Catégorie supprimée !');
+        } else {
+            $this->addFlash('danger', 'Token absent ou invalide !');
+        }
         return $this->redirectToRoute('admin_category_index');
     }
 }
